Re-factor resolvconf code to it's own class
Funcationally the same, but all the code related to DNS set and cleanup
is now in a class of it's own, which is dependency injected so that
someone could write a version to directy manipulate /etc/resolv.conf if
they are feeling masochistic.
