+++ nxbender/nx.py
@@ -5,7 +5,6 @@ from . import ppp
 import pyroute2
 import ipaddress
 import atexit
-import subprocess
 
 from requests.adapters import HTTPAdapter
 from requests.packages.urllib3.poolmanager import PoolManager
@@ -27,8 +26,14 @@ class FingerprintAdapter(HTTPAdapter):
             block=block, assert_fingerprint=self.fingerprint)
 
 class NXSession(object):
-    def __init__(self, options):
+    def __init__(self, options, dns_handler = None):
+        """Constructor
+
+        param options The parsed command line options to the program.
+        param dns_handler Object which will be used to apply DNS settings to the system.
+        """
         self.options = options
+        self.dns_handler = dns_handler
 
     def run(self):
         self.host = self.options.server + ':%d' % self.options.port
@@ -151,9 +156,10 @@ class NXSession(object):
 
         pppd = ppp.PPPSession(self.options, auth_key, routecallback=self.post_connect)
         pppd.run()
-        # remove DNS entries as they will have stopped working
-        if self.options.use_resolvconf:
-            self.remove_dns_resolvconf(self.device)
+        # reach here when PPPD exits
+        if self.dns_handler is not None:
+            # remove DNS entries as they will have stopped working
+            self.dns_handler.RemoveDns(self.device, self.srv_options)
 
     def post_connect(self, gateway:str, device:str):
         """Called after the PPP channel connects and obtains an IP address to complete setup.
@@ -165,8 +171,8 @@ class NXSession(object):
         # set IP routes
         self.setup_routes(gateway)
         # if needed, set DNS servers
-        if self.options.use_resolvconf:
-            self.set_dns_resolvconf(device)
+        if self.dns_handler is not None:
+            self.dns_handler.SetDns(device, self.srv_options)
 
     def setup_routes(self, gateway):
         """Called after the PPP channel connects and obtains an IP address to create routing table
@@ -180,66 +186,3 @@ class NXSession(object):
             ip.route("add", dst=dst, gateway=gateway)
 
         logging.info("Remote routing configured, VPN is up")
-
-    def set_dns_resolvconf(self, device:str):
-        """Called after the PPP channel connects to configure DNS using resolvconf.
-
-        param device The name of the local network device, e.g. ppp0
-        """
-        resolv_conf_lines = []
-        # header
-        resolv_conf_lines.append(f'# Generated by nxBender for {device}')
-
-        # nameserver lines
-        for dns in ['dns1', 'dns2', 'dns3']:
-            try:
-                resolv_conf_lines.append(f'nameserver {self.srv_options[dns]}')
-            except KeyError as err:
-                # ignore these, because the number of DNS servers returned will vary
-                pass
-
-        # search lines
-        # The server sends both "dnsSuffixes" and "dnsSuffix" with the same (single) value
-        # As I don't know how to parse the former, use the latter for now.
-        try:
-            domain = self.srv_options['dnsSuffix']
-            resolv_conf_lines.append(f'search {domain}')
-        except KeyError as err:
-                # ignore if not sent
-                pass
-
-        # pass the lot to resolvconf
-        # <data> | resolvconf -a ppp0
-        logging.debug("DNS settings to add:")
-        for line in resolv_conf_lines:
-            logging.debug(f'\t{line}')
-        try:
-            resolvconf = subprocess.run(['resolvconf', '-a'] + [device],
-                                         text = True,
-                                         input ='\n'.join(resolv_conf_lines),
-                                         check = True
-                                         )
-        except OSError as e:
-            logging.error("Unable to run resolvconf: %s" % e.strerror)
-        except subprocess.CalledProcessError as err:
-            logging.error(
-                f'resolvconf exited with status {err.returncode}: {err.stdout} {err.stderr}')
-
-
-    def remove_dns_resolvconf(self, device:str):
-        """Called after the PPP channel disconnects to remove DNS settings using resolvconf.
-
-        param device The name of the local network device, e.g. ppp0
-        """
-        data =""
-        try:
-            resolvconf = subprocess.run(['resolvconf', '-f', '-d'] + [device],
-                                         text = True,
-                                         input = data,
-                                         check = True
-                                         )
-        except OSError as e:
-            logging.error("Unable to run resolvconf: %s" % e.strerror)
-        except subprocess.CalledProcessError as err:
-            logging.error(
-                f'resolvconf exited with status {err.returncode}: {err.stdout} {err.stderr}')
